fix service discovery address notification and migration bugs

+++ dubbo-registry/dubbo-registry-api/src/main/java/org/apache/dubbo/registry/client/migration/DefaultMigrationAddressComparator.java
@@ -62,7 +62,7 @@ public class DefaultMigrationAddressComparator implements MigrationAddressCompar
             return false;
         }
 
-        if ((float) (newAddressSize / oldAddressSize) >= threshold) {
+        if (((float)newAddressSize / (float)oldAddressSize) >= threshold) {
             return true;
         }
         return false;
